requirements.txt
Pillow==9.4.0
transformers==4.26.1
torch==1.13.1
torchvision==1.13.1 
torchaudio==1.13.1
streamlit===1.19.0
sentencepiece==0.1.97 